metrics.default_metrics.tarantool.utils moved to metrics.utils
